assignment history update
