+++ docker-compose-api-only.yml
@@ -37,6 +37,7 @@ services:
       - CLIENTS_API_V2_URL=http://host.docker.internal:3333/v2/clients
       - TP_API_URL=http://host.docker.internal:3333/v1/TrackingPrograms
       - AGENDA_ITEM_API_URL=http://host.docker.internal:3333/v1/AgendaItems
+      - ASSIGNMENT_HISTORY_API_V2_URL=http://host.docker.internal:3333/v1/AssignmentHistory
       - PERSON_API_URL=http://host.docker.internal:3333/v3/persons
       - EMAIL_ENABLED=true
       - EMAIL_HOST=host.docker.internal
