+++ talentmap_api/fsbid/urls/client.py
@@ -9,7 +9,6 @@ urlpatterns = [
     url(r'^(?P<pk>[0-9]+)/suggestions/$', views.FSBidClientSuggestionsView.as_view(), name='FSBid-client-suggestions'),
     url(r'^export/$', views.FSBidClientCSVView.as_view(), name="FSBid-client_export"),
     url(r'^(?P<pk>[0-9]+)/$', views.FSBidClientView.as_view(), name='FSBid-client'),
-    url(r'^(?P<pk>[0-9]+)/assignments/$', views.FSBidClientAssignmentsHistoryView.as_view(), name='FSBid-client'),
     url(r'^$', views.FSBidClientListView.as_view(), name='FSBid-client_list'),
 ]
