+++ talentmap_api/settings.py
@@ -544,6 +544,7 @@ HRDATA_URL_EXTERNAL = get_delineated_environment_variable('HRDATA_URL_EXTERNAL',
 AVATAR_URL = get_delineated_environment_variable('AVATAR_URL', 'https://usdos.sharepoint.com/_layouts/15/userphoto.aspx')
 TP_API_URL = get_delineated_environment_variable('TP_API_URL', 'http://mock_fsbid:3333/v1/TrackingPrograms')
 AGENDA_ITEM_API_URL = get_delineated_environment_variable('AGENDA_ITEM_API_URL', 'http://mock_fsbid:3333/v1/AgendaItems')
+ASSIGNMENT_HISTORY_API_V2_URL = get_delineated_environment_variable('ASSIGNMENT_HISTORY_API_V2_URL', 'http://mock_fsbid:3333/v2/AssignmentHistory')
 PERSON_API_URL = get_delineated_environment_variable('PERSON_API_URL', 'http://mock_fsbid:3333/v3/persons')
 
 SAML_CONFIG_LOADER = 'talentmap_api.settings.config_settings_loader'
