+++ talentmap_api/urls.py
@@ -51,6 +51,7 @@ urlpatterns = [
     # Agenda
     url(r'^api/v1/fsbid/agenda/', include('talentmap_api.fsbid.urls.agenda')),
     url(r'^api/v1/fsbid/agenda_employees/', include('talentmap_api.fsbid.urls.agenda_employees')),
+    url(r'^api/v1/fsbid/assignment_history/', include('talentmap_api.fsbid.urls.assignment_history')),
 
     # Projected Vacancies
     url(r'^api/v1/projected_vacancy/', include('talentmap_api.projected_vacancies.urls.projected_vacancies')),
